+++ vscode/src/utils/FileSystem.spec.ts
@@ -1,7 +1,6 @@
 const _path = require('path');
 
 import { posixRelativePath, createFileSystem } from "./FileSystem";
-import { NoteSearcherConfigImpl } from "./NoteSearcherConfig";
 
 describe('FileSystem', () => {
   describe('posixRelativePath', () => {
@@ -29,19 +28,14 @@ describe('FileSystem', () => {
 
 const demoDir = _path.resolve(__dirname, '../../demo_dir');
 const readmePath = _path.resolve(demoDir, 'readme.md');
-const notIgnoredFilePath = _path.resolve(demoDir, 'not_ignored_stuff/not_ignored_file.md');
-const ignoredFilePath = _path.resolve(demoDir, 'ignored_stuff/ignored_file.md');
-const nestedNotIgnoredFilePath = _path.resolve(demoDir, 'subdir/ignored_stuff/not_ignored.md');
-const topNodeModulesFilePath = _path.resolve(demoDir, 'node_modules/about_node_modules.md');
-const nestedNodeModulesFilePath = _path.resolve(demoDir, 'subdir/node_modules/nested_node_modules.md');
 
 describe('FileSystem, demo dir', () => {
   const fs = createFileSystem();
 
   it('ignores specific file', () => {
-    const ignore = (path: string) => path === ignoredFilePath;
+    const ignore = (path: string) => path === readmePath;
     const allDemoDirFiles = Array.from(fs.allFilesUnderPath(demoDir, ignore));
-    expect(allDemoDirFiles).not.toContain(ignoredFilePath);
+    expect(allDemoDirFiles).not.toContain(readmePath);
   });
 
   it('ignores everything', () => {
@@ -51,54 +45,31 @@ describe('FileSystem, demo dir', () => {
   });
 });
 
-describe('FileSystem + config, demo dir, allFilesUnderPath', () => {
+describe('FileSystem, demo dir, allFilesUnderPath', () => {
   let allDemoDirFiles: string[];
 
   beforeEach(() => {
     const fs = createFileSystem();
-    const config = NoteSearcherConfigImpl.fromWorkspace(demoDir, fs);
-    allDemoDirFiles = Array.from(fs.allFilesUnderPath(demoDir, config.isIgnored));
+    allDemoDirFiles = Array.from(fs.allFilesUnderPath(demoDir));
   });
 
   it('contains readme', () => {
     expect(allDemoDirFiles).toContain(readmePath);
-  });
-
-  it('contains not_ignored_file', () => {
-    expect(allDemoDirFiles).toContain(notIgnoredFilePath);
-  });
-
-  it('contains nested not_ignored_file', () => {
-    expect(allDemoDirFiles).toContain(nestedNotIgnoredFilePath);
-  });
-
-  it('does not contain ignored_file' , () => {
-    expect(allDemoDirFiles).not.toContain(ignoredFilePath);
-  });
-
-  it('does not contain any node_modules' , () => {
-    expect(allDemoDirFiles).not.toContain(topNodeModulesFilePath);
-    expect(allDemoDirFiles).not.toContain(nestedNodeModulesFilePath);
   });
 });
 
 const demoSubDir = _path.resolve(__dirname, '../../demo_dir/subdir');
 const subFile = _path.resolve(__dirname, '../../demo_dir/subdir/my_sub_file.md');
 
-describe('FileSystem + config, demo subdir, allFilesUnderPath', () => {
+describe('FileSystem, demo subdir, allFilesUnderPath', () => {
   let allDemoDirFiles: string[];
 
   beforeEach(() => {
     const fs = createFileSystem();
-    const config = NoteSearcherConfigImpl.fromWorkspace(demoDir, fs);
-    allDemoDirFiles = Array.from(fs.allFilesUnderPath(demoSubDir, config.isIgnored));
+    allDemoDirFiles = Array.from(fs.allFilesUnderPath(demoSubDir));
   });
 
   it('contains subfile', () => {
     expect(allDemoDirFiles).toContain(subFile);
-  });
-
-  it('does not contain nested_node_modules', () => {
-    expect(allDemoDirFiles).not.toContain(nestedNodeModulesFilePath);
   });
 });